refactor(google): restructure Google API contracts and imports
- consolidate Google API interface and session display type definitions
- update imports to reflect new file structure for better organization
- enhance clarity and maintainability of Google service contracts
